storage-node-v2: Refactor the code.
- move code to the separate files
